Created a separate ApolloContainer and wrapped entire app around it
